Added support for rg350

+++ make_release_packages.sh
@@ -7,6 +7,7 @@ echo "<head>" >> index.htm
 echo "</head>" >> index.htm
 echo "<body>" >> index.htm
 TIME=`date -u +"%d.%m.%Y %R"`
+echo "Version $VERSION" >> index.htm
 echo "Updated at the $TIME." >> index.htm
 echo "<h1>$PROGRAM download links:</h1>" >> index.htm
 echo "<?php" > symlink.php
@@ -40,11 +41,17 @@ do
 					mv "$PROGRAM.opk" ../..
 					echo "<a href=$PROGRAM.opk type=\"application/x-opk+squashfs\">$NAME</a></br>" >> ../../index.htm
 				else
-					zip -r "$PROGRAM-$NAME-$VERSION.zip" * > /dev/null
-					mv "$PROGRAM-$NAME-$VERSION.zip" ../..
-					echo "<a href=$PROGRAM-$NAME-$VERSION.zip>$NAME</a></br>" >> ../../index.htm
-					echo "unlink('$PROGRAM-$NAME.zip');" >> ../../symlink.php
-					echo "symlink('$PROGRAM-$NAME-$VERSION.zip', '$PROGRAM-$NAME.zip');" >> ../../symlink.php
+					if [ $NAME = "rg350" ]; then
+						mksquashfs * "$PROGRAM-rg350.opk" -all-root -noappend -no-exports -no-xattrs
+						mv "$PROGRAM-rg350.opk" ../..
+						echo "<a href=$PROGRAM-rg350.opk type=\"application/x-opk+squashfs\">$NAME</a></br>" >> ../../index.htm
+					else
+						zip -r "$PROGRAM-$NAME-$VERSION.zip" * > /dev/null
+						mv "$PROGRAM-$NAME-$VERSION.zip" ../..
+						echo "<a href=$PROGRAM-$NAME-$VERSION.zip>$NAME</a></br>" >> ../../index.htm
+						echo "unlink('$PROGRAM-$NAME.zip');" >> ../../symlink.php
+						echo "symlink('$PROGRAM-$NAME-$VERSION.zip', '$PROGRAM-$NAME.zip');" >> ../../symlink.php
+					fi
 				fi
 			fi
 		fi
